Refactor graph review tool into review_gui package.
Replace monolithic review_graph.py and review_app.py with server, streamlit_ui,
and a render subpackage; node thumbnails remain placeholders until SimApp lands.
