+++ isaaclab_arena_examples/agentic_environment_generation/review_gui/render/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2026, The Isaac Lab Arena Project Developers (https://github.com/isaac-sim/IsaacLab-Arena/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""HTML rendering backend for the initial-graph review dashboard."""
